Add BotUploader interface & local implementation
This class models everything needed for creating a new bot. It should ensure the bot can be executed correctly by validating it first, then uploading to an storage service & updating the db with the new bot.

A local implementation has been done, where these steps are done secuentially in the client. However, another implementation could be calling via REST some API that does these steps and returns the results to the client.